폴더위치 변경 및 라우터를 하는 tsx 파일 따로 생성 (수정 중)
