+++ src/pages/kojaem/kirsh/Kirsh.tsx
